extract solver module
